Rename RepositoryTupleKeyListNameGenerator (#31141)
